Refactor + comments code
